Moved subscription validation into subscriber service.
